Fix resolution issue in cycle generator
git-svn-id: svn://svn.icculus.org/obdgpslogger/trunk@532 ec62d176-4a65-0410-9229-e4683bb12484

+++ src/sim/generators/cycle/gen_cycle.c
@@ -109,7 +109,7 @@ int cycle_simgen_getvalue(void *gen, unsigned int mode, unsigned int PID, unsign
 	gettimeofday(&newtime, NULL);
 
 	float dt = (newtime.tv_sec - g->firsttime.tv_sec)
-				+ (((long)newtime.tv_usec - (long)g->firsttime.tv_usec) / US_TO_SEC);
+				+ ((float)((long)newtime.tv_usec - (long)g->firsttime.tv_usec) / (float)US_TO_SEC);
 
 	if(dt < 0) {
 		printf("Cycle dt<0! dt: %f , newtime: %li %li , firsttime: %li %li\n",
